Reset DB when --flush_db is specified
The previous code assumes that the sampler (rank=0) is launched at
first. However, it turned out that in the test code
(tests/sampler/test_sampler.py) the sampler is launched after the master
since it took some time to spawn a new process. As a result, flushing DB
if rank == 0 results in deleting a new trigger key that the master
already set. Instead, adding --flush_db flag and if and only if
--flush_db is given, the constructor performs flushing DB. Set
--flush_db appropriately is programmers responsibility.

I don't add any modification to the test code since there is no need to
flush DB.